Refactored items => fields + added experimental form
